pyproject.toml
[project]
name = "week6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.7",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-stubs>=5.2.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
